% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Allfuncs.R
\name{combinePaths}
\alias{combinePaths}
\title{combines binary pathway matricies into one, rownames are matched by name}
\usage{
combinePaths(...)
}
\arguments{
\item{...}{binary pathway matricies}
}
\description{
combines binary pathway matricies into one, rownames are matched by name
}
