% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Allfuncs.R
\name{commonRows}
\alias{commonRows}
\title{returns the row names in common}
\usage{
commonRows(data1, data2)
}
\arguments{
\item{data1}{One matrix with gene rownames}

\item{data2}{Another matrix with gene rownames}
}
\description{
returns the row names in common
}
