% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Allfuncs.R
\name{crossVal}
\alias{crossVal}
\title{crossVal}
\usage{
crossVal(plierRes, data, priorMat, priorMatcv)
}
\arguments{
\item{priorMat}{the real prior info matrix}

\item{priorMatcv}{the zeroed-out prior info matrix used for PLIER
computations}
}
\description{
crossVal
}
\keyword{internal}
