% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Allfuncs.R
\name{DataSmooth}
\alias{DataSmooth}
\title{SVD based smoothing for single cell RNAseq data}
\usage{
DataSmooth(svdres, k)
}
\arguments{
\item{svdres}{svd result}

\item{k}{number of components to use}
}
\description{
SVD based smoothing for single cell RNAseq data
}
