% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Allfuncs.R
\name{getAUC}
\alias{getAUC}
\title{getAUC}
\usage{
getAUC(plierRes, data, priorMat)
}
\arguments{
\item{plierRes}{current PLIER result}

\item{data}{the input data}

\item{priorMat}{the  prior info matrix}
}
\description{
getAUC
}
\keyword{internal}
