% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Allfuncs.R
\name{getEnrichmentVals}
\alias{getEnrichmentVals}
\title{getEnrichmentVals}
\usage{
getEnrichmentVals(
  plierRes,
  pathwayMat,
  ngenes = 50,
  auc.cutoff = 0.7,
  fdr.cutoff = 0.01
)
}
\description{
getEnrichmentVals
}
\keyword{internal}
