% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Allfuncs.R
\name{plotMat}
\alias{plotMat}
\title{generic function to plot the non-zero elements of a sparse matrix}
\usage{
plotMat(matrix, scale = T, trim.names = 50, cutoff = NULL,
  col.scale = NULL, ...)
}
\arguments{
\item{matrix}{the input matrix}

\item{scale}{rescale the columns to max=1}

\item{trim.names}{the maximum length of row and column lables}

\item{cutoff}{Sets values (both positive and negative) bellow this number to 0}

\item{col.scale}{custom color scale}

\item{...}{additional argumend to be passed to pheatmap}
}
\description{
generic function to plot the non-zero elements of a sparse matrix
}
