% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Allfuncs.R
\name{plotTopZ}
\alias{plotTopZ}
\title{visualize the top genes contributing to the LVs}
\usage{
plotTopZ(plierRes, data, priorInfo, top = 10, index = NULL)
}
\arguments{
\item{plierRes}{the result returned by PLIER}

\item{data}{the data to be displayed in a heatmap, typically the z-scored input data (or some subset thereof)}

\item{priorInfo}{the same gene by geneset binary matrix that was used to run PLIER}

\item{top}{the top number of genes to use}

\item{index}{the subset of LVs to display}
}
\description{
visualize the top genes contributing to the LVs
}
