% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Allfuncs.R
\name{projectPLIER}
\alias{projectPLIER}
\title{project a new dataset onto PLIER LVs}
\usage{
projectPLIER(PLIERres, newdata)
}
\arguments{
\item{PLIERres}{output of PLIER or simpleDecomp}

\item{newdata}{new data to project into LV space}
}
\description{
project a new dataset onto PLIER LVs
}
\examples{
data("dataWholeBlood")
res=simpleDecomp(datain<-tscale(dataWholeBlood), 20)
newB=projectPLIER(res, datain)
}
