% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Allfuncs.R
\name{rotateSVD}
\alias{rotateSVD}
\title{rotate SVD to maximize the L1  of positive U values}
\usage{
rotateSVD(svdres)
}
\arguments{
\item{svdres}{a result from a call to svd or rsvd}
}
\description{
rotate SVD to maximize the L1  of positive U values
}
\keyword{internal}
