% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Allfuncs.R
\name{simpleDecomp}
\alias{simpleDecomp}
\title{Main PLIER function}
\usage{
simpleDecomp(
  Y,
  k,
  svdres = NULL,
  L1 = NULL,
  L2 = NULL,
  max.iter = 200,
  tol = 5e-06,
  trace = F,
  rseed = NULL,
  B = NULL,
  scale = 1,
  adaptive.frac = 0.05,
  adaptive.iter = 30
)
}
\arguments{
\item{k}{The number of latent variables to return, leave as NULL to be set automatically using the num.pc "elbow" method}

\item{svdres}{Pre-computed result of the svd decomposition for data.}

\item{L1}{L1 constant, leave as NULL to automatically select a value.}

\item{L2}{L2 constant, leave as NULL to automatically select a value.}

\item{max.iter}{maximum number of iterations, default is 200.}

\item{tol}{tolerance for relative B change, default is 5e-6.}

\item{trace}{print out trace info, default is false.}

\item{rseed}{seed for RSVD call. If svdres us given this has no effect.}

\item{adaptive.frac}{the probability cutoff to use to determine the positive cutoff for loading. Smaller values will return sparser models.}

\item{adaptive.iter}{the first iteration where adaptive constants are computed, before they are all 0.}

\item{data}{the data to be processed with genes in rows and samples in columns. Should be z-scored.}

\item{scaleL}{scale both L1 and L2 by this amount from the default setting. Values less then 1 imply weaker regularization.}
}
\description{
Main PLIER function
}
