% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Allfuncs.R
\name{solveU}
\alias{solveU}
\title{Solves for the U coefficients making efficient utilizatoin of the lasso path}
\usage{
solveU(Z, Chat, priorMat, penalty.factor, pathwaySelection = "fast",
  glm_alpha = 0.9, maxPath = 10, target.frac = 0.7, L3 = NULL)
}
\arguments{
\item{Z}{current Z estimate}

\item{Chat}{the inverse of the C matrix}

\item{priorMat}{the prior pathway or C matrix}

\item{penalty.factor}{Penalties for different pathways, must have size ncol(priorMat).}

\item{pathwaySelection}{Method to use for pathway selection.}

\item{glm_alpha}{The elsatic net alpha parameter}

\item{maxPath}{The maximum number of pathways to consider}

\item{target.frac}{The target fraction on non-zero columns of}

\item{L3}{Solve with a given L3, no search}
}
\description{
Solves for the U coefficients making efficient utilizatoin of the lasso path
}
\keyword{internal}
