% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Allfuncs.R
\name{mapPathway}
\alias{mapPathway}
\title{mapPathway}
\usage{
mapPathway(pathway, mapping)
}
\arguments{
\item{pathway}{pathway matrix}

\item{mapping}{Gene name map. A single column data.frame or matrix with map-from
in row names and map-to in the first column}
}
\description{
Rename pathway matrix gene names. Useful for species conversion
}
