% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Allfuncs.R
\name{nameB}
\alias{nameB}
\title{names latent variables according to their pathway useage (if any)}
\usage{
nameB(plierRes, top = 1, fdr.cutoff = 0.01, use = c("coef", "AUC"))
}
\arguments{
\item{plierRes}{A PLIER result}

\item{top}{The number of pathway to use. Only the top pathway (one with the largest coefficient) is used by default}

\item{fdr.cutoff}{The cross-validation significance cutoff for a pathway to be considered for naming. If no pathways satisfy the cutoff the raw coefficients are used.}

\item{Choice}{of coef or AUC, whether LVs are named based on U coefficients or AUCs. Defualt: coef.}
}
\description{
names latent variables according to their pathway useage (if any)
}
